Make assignmentTabs more general tabs component
